+++ client/src/pages/reservationManagement/ReservationGroupOverview.jsx
@@ -1,7 +1,11 @@
 import React, { useEffect, useState } from "react";
 import { fetchReservationGroupsById } from "../../store/actions/ReservationGroupActions";
 import { deleteReservationGroup } from "../../store/actions/ReservationGroupActions";
-import { editReservationGroup } from "../../store/actions/ReservationGroupActions";
+import {
+  editReservationGroup,
+  fetchReservationItemByGroupId,
+} from "../../store/actions/ReservationGroupActions";
+
 import { DeleteConfirmModel } from "../../components/DeleteConfirmModel";
 import TitleActionBar from "../../components/TitleActionsBar";
 import { useLocation, useNavigate } from "react-router-dom";
@@ -18,7 +22,10 @@ const ManageReservationGroups = () => {
   const fetchReservationGroupData = useSelector(
     (state) => state.getReservationGroupById.fetchReservationGroupId
   );
-  
+  const fetchReservationItemByGroupData = useSelector(
+    (state) => state.fetchReservationItemByGroup.fetchReservationItemByGroupFlag
+  );
+
   const [recordId, setRecordId] = useState("");
   const [groupId, setGroupId] = useState("");
   const [groupName, setGroupName] = useState("");
@@ -28,6 +35,7 @@ const ManageReservationGroups = () => {
   const [isSaveDisable, setIsSaveDisable] = useState(true);
   const [isDeleteDisable, setIsDeleteDisable] = useState(false);
   const [showConfirmation, setShowConfirmation] = useState(false);
+  const [itemsExist, setItemsExist] = useState(false);
   const [count, setCount] = useState(0);
   const searchParams = new URLSearchParams(useLocation().search);
   const data = searchParams.get("data");
@@ -37,8 +45,21 @@ const ManageReservationGroups = () => {
   useEffect(() => {
     if (paramData && recordId) {
       dispatch(fetchReservationGroupsById(recordId));
+      dispatch(fetchReservationItemByGroupId(recordId));
     }
   }, [dispatch, recordId]);
+
+  useEffect(() => {
+    if (fetchReservationItemByGroupData) {
+      console.log(
+        "fetchReservationItemByGroupData",
+        fetchReservationItemByGroupData
+      );
+      setItemsExist(fetchReservationItemByGroupData);
+    } else {
+      setItemsExist(fetchReservationItemByGroupData);
+    }
+  }, [fetchReservationItemByGroupData]);
 
   const fetchData = () => {
     if (fetchReservationGroupData) {
@@ -91,7 +112,7 @@ const ManageReservationGroups = () => {
           groupId: groupId,
           groupName: groupName,
         };
-        
+
         dispatch(editReservationGroup(recordId, formData));
         handleNavigate();
         toast.success("Data saved successfully");
@@ -106,9 +127,13 @@ const ManageReservationGroups = () => {
   const confirmDelete = async () => {
     try {
       if (paramData && paramData.id) {
-        dispatch(deleteReservationGroup(paramData.id));
-        toast.success("Data deleted successfully");
-        handleNavigate();
+        if (!itemsExist) {
+          dispatch(deleteReservationGroup(paramData.id));
+          toast.success("Data deleted successfully");
+          handleNavigate();
+        } else {
+          toast.error("Cannot delete group with items");
+        }
       } else {
         toast.error("Cannot delete. ID is undefined.");
       }
